+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.3] - 2023-09-12
+
 ### Added
 
-- Support for nested popovers
+- Support for nested popovers ([#43](https://github.com/Ambiki/impulse_view_components/pull/43))
 
 ## [0.2.2] - 2023-09-11
 
@@ -79,7 +81,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Everything!
 
-[unreleased]: https://github.com/Ambiki/impulse_view_components/compare/v0.2.2...HEAD
+[unreleased]: https://github.com/Ambiki/impulse_view_components/compare/v0.2.3...HEAD
+[0.2.3]: https://github.com/Ambiki/impulse_view_components/compare/v0.2.2...v0.2.3
 [0.2.2]: https://github.com/Ambiki/impulse_view_components/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/Ambiki/impulse_view_components/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/Ambiki/impulse_view_components/compare/v0.1.3...v0.2.0